Installer for LAMP with Composer and Docker

+++ vhost.sh
@@ -5,23 +5,34 @@
 ### if you use larave, path should be to use /public
 ### installer.sh amirhome "Amir/amirhome/public" real path is /var/www/vhosts/..
 
-echo "127.0.0.1 $1.loc" | sudo tee -a /etc/hosts;
+
+VPATH=$1
+if [ -z $1 ]
+then
+	echo "ServerName is Required"
+	exit
+fi
+
+
+if grep -q "127.0.0.1 $1.loc" "/etc/hosts"; then
+  echo "127.0.0.1 $1.loc" | sudo tee -a /etc/hosts
+fi
 
 sudo rm -rf /etc/apache2/sites-enabled/$1.loc.conf
 
-path=$1
+
 if [ ! -z "$2" ]
 then
-      path=$2
+    VPATH="$2";
 fi
 
 
 sudo tee -a /etc/apache2/sites-enabled/$1.loc.conf > /dev/null <<EOT
 <VirtualHost *:80>
 	ServerName $1.loc
-	DocumentRoot /var/www/vhosts/$path
+	DocumentRoot $HOME/Codes/vhosts/$VPATH
 	SetEnv MAGE_MODE developer
-	<Directory "/var/www/vhosts/$path/">
+	<Directory "$HOME/Codes/vhosts/$VPATH/">
 		AllowOverride All
 		Order allow,deny
 		Allow from all
@@ -29,4 +40,7 @@ sudo tee -a /etc/apache2/sites-enabled/$1.loc.conf > /dev/null <<EOT
 </VirtualHost>
 EOT
 
+
 sudo systemctl restart apache2;
+
+echo "Enjoying http://$1.loc , call from $HOME/Codes/$VPATH"
